Fixed: The program didn't start unless the program folder was explicitly specified using the -userdir parameter
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@8927 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/netmisc.h
@@ -141,7 +141,7 @@ void BESwapObject (CObject *obj);
 	memcpy ((ubyte *)(_extraGameInfo), data, sizeof(tExtraGameInfo)); AddPlayerLoadout (false)
 
 #define SendInternetXMLGameInfoPacket(xmlGameInfo, server, node) \
-	IPXSendInternetPacketData((ubyte *) xmlGameInfo, strlen(xmlGameInfo) + 1, server, node)
+	IPXSendInternetPacketData((ubyte *) xmlGameInfo, (int) strlen(xmlGameInfo) + 1, server, node)
 
 #define SendMissingObjFramesPacket(server, node, netAddress) \
 	IPXSendPacketData((ubyte *) &networkData.sync [0].objs.missingFrames, sizeof(tMissingObjFrames), server, node, netAddress)
